Struct defindex_factory::DeFindexFactoryClient

pub struct DeFindexFactoryClient<'a> {
    pub env: Env,
    pub address: Address,
    /* private fields */
}

DeFindexFactoryClient is a client for calling the contract defined in “DeFindexFactory”.

Fields
env: Envaddress: Address
Implementations

impl<'a> DeFindexFactoryClient<'a>

pub fn new(env: &Env, address: &Address) -> Self

impl<'a> DeFindexFactoryClient<'a>

pub fn initialize( &self, admin: &Address, defindex_receiver: &Address, defindex_fee: &u32, vault_wasm_hash: &BytesN<32>, )

Initializes the factory contract with the given parameters.

Arguments

e - The environment in which the contract is running.

admin - The address of the contract administrator, who can manage settings.

defindex_receiver - The default address designated to receive a portion of fees.

defindex_fee - The initial annual fee rate (in basis points).

vault_wasm_hash - The hash of the DeFindex Vault’s WASM file for deploying new vaults.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, otherwise an error.

pub fn try_initialize( &self, admin: &Address, defindex_receiver: &Address, defindex_fee: &u32, vault_wasm_hash: &BytesN<32>, ) -> Result<Result<(), <() as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Initializes the factory contract with the given parameters.

Arguments

e - The environment in which the contract is running.

admin - The address of the contract administrator, who can manage settings.

defindex_receiver - The default address designated to receive a portion of fees.

defindex_fee - The initial annual fee rate (in basis points).

vault_wasm_hash - The hash of the DeFindex Vault’s WASM file for deploying new vaults.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, otherwise an error.

pub fn create_defindex_vault( &self, emergency_manager: &Address, fee_receiver: &Address, vault_fee: &u32, vault_name: &String, vault_symbol: &String, manager: &Address, assets: &Vec<AssetStrategySet>, salt: &BytesN<32>, ) -> Address

Creates a new DeFindex Vault with specified parameters.

Arguments

e - The environment in which the contract is running.

emergency_manager - The address assigned emergency control over the vault.

fee_receiver - The address designated to receive fees from the vault.

vault_fee - The percentage share of fees allocated to the vault’s fee receiver.

manager - The address assigned as the vault manager.

assets - A vector of AssetStrategySet structs that define the assets managed by the vault.

salt - A salt used for ensuring unique addresses for each deployed vault.

Returns

Result<Address, FactoryError> - Returns the address of the new vault, or an error if unsuccessful.

pub fn try_create_defindex_vault( &self, emergency_manager: &Address, fee_receiver: &Address, vault_fee: &u32, vault_name: &String, vault_symbol: &String, manager: &Address, assets: &Vec<AssetStrategySet>, salt: &BytesN<32>, ) -> Result<Result<Address, <Address as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Creates a new DeFindex Vault with specified parameters.

Arguments

e - The environment in which the contract is running.

emergency_manager - The address assigned emergency control over the vault.

fee_receiver - The address designated to receive fees from the vault.

vault_fee - The percentage share of fees allocated to the vault’s fee receiver.

manager - The address assigned as the vault manager.

assets - A vector of AssetStrategySet structs that define the assets managed by the vault.

salt - A salt used for ensuring unique addresses for each deployed vault.

Returns

Result<Address, FactoryError> - Returns the address of the new vault, or an error if unsuccessful.

pub fn create_defindex_vault_deposit( &self, caller: &Address, emergency_manager: &Address, fee_receiver: &Address, vault_fee: &u32, vault_name: &String, vault_symbol: &String, manager: &Address, assets: &Vec<AssetStrategySet>, amounts: &Vec<i128>, salt: &BytesN<32>, ) -> Address

Creates a new DeFindex Vault with specified parameters and makes the first deposit to set ratios.

Arguments

e - The environment in which the contract is running.

emergency_manager - The address assigned emergency control over the vault.

fee_receiver - The address designated to receive fees from the vault.

vault_fee - The percentage share of fees allocated to the vault’s fee receiver.

vault_name - The name of the vault.

vault_symbol - The symbol of the vault.

manager - The address assigned as the vault manager.

assets - A vector of AssetStrategySet structs that define the assets managed by the vault.

amounts - A vector of AssetAmounts structs that define the initial deposit amounts.

salt - A salt used for ensuring unique addresses for each deployed vault.

Returns

Result<Address, FactoryError> - Returns the address of the new vault, or an error if unsuccessful.

pub fn try_create_defindex_vault_deposit( &self, caller: &Address, emergency_manager: &Address, fee_receiver: &Address, vault_fee: &u32, vault_name: &String, vault_symbol: &String, manager: &Address, assets: &Vec<AssetStrategySet>, amounts: &Vec<i128>, salt: &BytesN<32>, ) -> Result<Result<Address, <Address as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Creates a new DeFindex Vault with specified parameters and makes the first deposit to set ratios.

Arguments

e - The environment in which the contract is running.

emergency_manager - The address assigned emergency control over the vault.

fee_receiver - The address designated to receive fees from the vault.

vault_fee - The percentage share of fees allocated to the vault’s fee receiver.

vault_name - The name of the vault.

vault_symbol - The symbol of the vault.

manager - The address assigned as the vault manager.

assets - A vector of AssetStrategySet structs that define the assets managed by the vault.

amounts - A vector of AssetAmounts structs that define the initial deposit amounts.

salt - A salt used for ensuring unique addresses for each deployed vault.

Returns

Result<Address, FactoryError> - Returns the address of the new vault, or an error if unsuccessful.

pub fn set_new_admin(&self, new_admin: &Address)

Sets a new admin address.

Arguments

e - The environment in which the contract is running.

new_admin - The new administrator’s address.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

pub fn try_set_new_admin( &self, new_admin: &Address, ) -> Result<Result<(), <() as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Sets a new admin address.

Arguments

e - The environment in which the contract is running.

new_admin - The new administrator’s address.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

pub fn set_defindex_receiver(&self, new_fee_receiver: &Address)

Updates the default receiver address for the DeFindex portion of fees.

Arguments

e - The environment in which the contract is running.

new_fee_receiver - The address of the new fee receiver.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

pub fn try_set_defindex_receiver( &self, new_fee_receiver: &Address, ) -> Result<Result<(), <() as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Updates the default receiver address for the DeFindex portion of fees.

Arguments

e - The environment in which the contract is running.

new_fee_receiver - The address of the new fee receiver.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

pub fn set_defindex_fee(&self, defindex_fee: &u32)

Updates the default fee rate for new vaults.

Arguments

e - The environment in which the contract is running.

new_fee_rate - The new annual fee rate in basis points.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

pub fn try_set_defindex_fee( &self, defindex_fee: &u32, ) -> Result<Result<(), <() as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Updates the default fee rate for new vaults.

Arguments

e - The environment in which the contract is running.

new_fee_rate - The new annual fee rate in basis points.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

pub fn admin(&self) -> Address

Retrieves the current admin’s address.

Arguments

e - The environment in which the contract is running.

Returns

Result<Address, FactoryError> - Returns the admin’s address or an error if not found.

pub fn try_admin( &self, ) -> Result<Result<Address, <Address as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Retrieves the current admin’s address.

Arguments

e - The environment in which the contract is running.

Returns

Result<Address, FactoryError> - Returns the admin’s address or an error if not found.

pub fn defindex_receiver(&self) -> Address

Retrieves the current DeFindex receiver’s address.

Arguments

e - The environment in which the contract is running.

Returns

Result<Address, FactoryError> - Returns the DeFindex receiver’s address or an error if not found.

pub fn try_defindex_receiver( &self, ) -> Result<Result<Address, <Address as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Retrieves the current DeFindex receiver’s address.

Arguments

e - The environment in which the contract is running.

Returns

Result<Address, FactoryError> - Returns the DeFindex receiver’s address or an error if not found.

pub fn deployed_defindexes(&self) -> Map<u32, Address>

Retrieves a map of all deployed DeFindex vaults.

Arguments

e - The environment in which the contract is running.

Returns

Result<Map<u32, Address>, FactoryError> - Returns a map with vault identifiers and addresses or an error if retrieval fails.

pub fn try_deployed_defindexes( &self, ) -> Result<Result<Map<u32, Address>, <Map<u32, Address> as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Retrieves a map of all deployed DeFindex vaults.

Arguments

e - The environment in which the contract is running.

Returns

Result<Map<u32, Address>, FactoryError> - Returns a map with vault identifiers and addresses or an error if retrieval fails.

pub fn defindex_fee(&self) -> u32

Retrieves the current fee rate.

Arguments

e - The environment in which the contract is running.

Returns

Result<u32, FactoryError> - Returns the fee rate in basis points or an error if not found.

pub fn try_defindex_fee( &self, ) -> Result<Result<u32, <u32 as TryFromVal<Env, Val>>::Error>, Result<FactoryError, InvokeError>>

Retrieves the current fee rate.

Arguments

e - The environment in which the contract is running.

Returns

Result<u32, FactoryError> - Returns the fee rate in basis points or an error if not found.