Trait defindex_factory::FactoryTrait

pub trait FactoryTrait {
    // Required methods
    fn initialize(
        e: Env,
        admin: Address,
        defindex_receiver: Address,
        fee_rate: u32,
        defindex_wasm_hash: BytesN<32>
    ) -> Result<(), FactoryError>;
    fn create_defindex_vault(
        e: Env,
        emergency_manager: Address,
        fee_receiver: Address,
        vault_share: u32,
        manager: Address,
        assets: Vec<AssetAllocation>,
        salt: BytesN<32>
    ) -> Result<Address, FactoryError>;
    fn set_new_admin(e: Env, new_admin: Address) -> Result<(), FactoryError>;
    fn set_defindex_receiver(
        e: Env,
        new_fee_receiver: Address
    ) -> Result<(), FactoryError>;
    fn set_fee_rate(e: Env, new_fee_rate: u32) -> Result<(), FactoryError>;
    fn admin(e: Env) -> Result<Address, FactoryError>;
    fn defindex_receiver(e: Env) -> Result<Address, FactoryError>;
    fn deployed_defindexes(e: Env) -> Result<Map<u32, Address>, FactoryError>;
    fn fee_rate(e: Env) -> Result<u32, FactoryError>;
}

Required Methods

fn initialize( e: Env, admin: Address, defindex_receiver: Address, fee_rate: u32, defindex_wasm_hash: BytesN<32> ) -> Result<(), FactoryError>

Initializes the factory contract with the given parameters.

Arguments

e - The environment in which the contract is running.

admin - The address of the contract administrator, who can manage settings.

defindex_receiver - The default address designated to receive a portion of fees.

fee_rate - The initial annual fee rate (in basis points).

defindex_wasm_hash - The hash of the DeFindex Vault’s WASM file for deploying new vaults.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, otherwise an error.

fn create_defindex_vault( e: Env, emergency_manager: Address, fee_receiver: Address, vault_share: u32, manager: Address, assets: Vec<AssetAllocation>, salt: BytesN<32> ) -> Result<Address, FactoryError>

Creates a new DeFindex Vault with specified parameters.

Arguments

e - The environment in which the contract is running.

emergency_manager - The address assigned emergency control over the vault.

fee_receiver - The address designated to receive fees from the vault.

vault_share - The percentage share of fees allocated to the vault’s fee receiver.

manager - The address assigned as the vault manager.

assets - A vector of AssetAllocation structs that define the assets managed by the vault.

salt - A salt used for ensuring unique addresses for each deployed vault.

Returns

Result<Address, FactoryError> - Returns the address of the new vault, or an error if unsuccessful.

fn set_new_admin(e: Env, new_admin: Address) -> Result<(), FactoryError>

Sets a new admin address.

Arguments

e - The environment in which the contract is running.

new_admin - The new administrator’s address.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

fn set_defindex_receiver( e: Env, new_fee_receiver: Address ) -> Result<(), FactoryError>

Updates the default receiver address for the DeFindex portion of fees.

Arguments

e - The environment in which the contract is running.

new_fee_receiver - The address of the new fee receiver.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

fn set_fee_rate(e: Env, new_fee_rate: u32) -> Result<(), FactoryError>

Updates the default fee rate for new vaults.

Arguments

e - The environment in which the contract is running.

new_fee_rate - The new annual fee rate in basis points.

Returns

Result<(), FactoryError> - Returns Ok(()) if successful, or an error if not authorized.

fn admin(e: Env) -> Result<Address, FactoryError>

Retrieves the current admin’s address.

Arguments

e - The environment in which the contract is running.

Returns

Result<Address, FactoryError> - Returns the admin’s address or an error if not found.

fn defindex_receiver(e: Env) -> Result<Address, FactoryError>

Retrieves the current DeFindex receiver’s address.

Arguments

e - The environment in which the contract is running.

Returns

Result<Address, FactoryError> - Returns the DeFindex receiver’s address or an error if not found.

fn deployed_defindexes(e: Env) -> Result<Map<u32, Address>, FactoryError>

Retrieves a map of all deployed DeFindex vaults.

Arguments

e - The environment in which the contract is running.

Returns

Result<Map<u32, Address>, FactoryError> - Returns a map with vault identifiers and addresses or an error if retrieval fails.

fn fee_rate(e: Env) -> Result<u32, FactoryError>

Retrieves the current fee rate.

Arguments

e - The environment in which the contract is running.

Returns

Result<u32, FactoryError> - Returns the fee rate in basis points or an error if not found.

Object Safety

This trait is not object safe.